organized common card component files
